back/requirements.txt
Flask~=3.0.2
python-dotenv~=1.0.1
SQLAlchemy~=2.0.29
psycopg2-binary~=2.9.1
flask_sqlalchemy
flask_cors